requirements.txt
livekit-agents
livekit-plugins-groq
livekit-plugins-elevenlabs
livekit-plugins-silero
python-dotenv
httpx